requirements.txt
asgiref==3.5.0
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==0.5.0
Django==3.2.12
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
PyMySQL==1.0.2
pyTelegramBotAPI==4.7.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
whitenoise==6.0.0
